Modelo de almacen y agregar productos
Se añade comportamiento para que los almacenes puedan tener la opcion de agregar elementos y en el menu se proporciona al usuario laas instrucciones para hacerlo

+++ src/market/admin/Client.java
@@ -10,9 +10,9 @@ public class Client {
 
     private class Purchase implements Comparable{
         private Lista<Product> shoppingCart;
-        private int sutotal;
-        private int total;
-        private int iva;
+        private double sutotal;
+        private double total;
+        private double iva;
 
         @Override
         public String toString() {
